create types.ts
